real type support: infinite precision real numbers
Adds lexing, parsing, typechecking and builtins

Builtins are currently unimplemented so this allows us
to load SAIL code that mentions real and the interpreter
should be able to run instructions that do not use real
but not to run those that do use real.

Also, the Lem and OCaml backends will not be able to generate
legal code if you use real.