Normalize vector start indices
In the Lem shallow embedding, vectors now always start at index 0, if the index
direction is increasing, and at index (len - 1), if the direction is
decreasing.  This is meant to simplify proofs when exporting to theorem
provers.

Explicit non-default start indices are therefore no longer necessary, and have
been removed from the vector representation in sail_values.lem.  The Sail
library and the pretty printer for the Lem shallow embedding have been adapted.