Improvements to implicit type casting
Added a new feature for implicit casts - now allowable implicit casts
can be specified by the user via a valspec such as

val cast forall Type 'a, Type 'b. 'a -> 'b effect pure cast_anything

with a new AST constructor to represent this as VS_cast_spec. This
constructor is removed and replaced with the standard val spec by the
old typechecker for backwards compatability, so it's only used by the
new typechecker, and won't appear in the ast once it reaches the
backends.

Also added Num as a synonym for the Nat kind in the parser, via the
confusingly named NatNum token (Num by itself was already taken for a
numeric constant).