Fix a bug with constraint generation in flow typing.
This bug manifested as the ARM example elf executable printing the
wrong characters... but otherwise doing not failing and exiting
cleanly. It didn't trigger the test suite at all. I tracked it down to
this line using git-bisect, and while returning nc_true is a bit
suspect I'm still not 100% sure how this caused such a subtle and
annoying bug in the generated ocaml code - it took several hours to
track down the breakage to this line.