Compile assertions into OCaml
and_bool and or_bool now are treated specially in the ocaml backend,
so that they have the correct short-circuiting behaviour. This is
required so that assertions don't fail for the ARM spec for predicates
that shouldn't be tested in certain circumstances, for example things like:

IsAArch32() && AArch32_specific_predicate

Also fixed an issue in the sail library for ocaml where greater than
or equal to was being mapped to greater than.