Split rewriter into separate rewriting library and rewrite passes
As discussed previously, we wanted to start refactoring the re-writer
to make it a bit less monolithic, and in the future potentially break
it into separate files for backend-specific rewrites and stuff.

- rewriter.ml now contains the generic re-writing code

- rewrites.ml contains the rewriting passes themselves

It would be nice if the generic rewriting code didn't depend on the
typechecker, because then it could be used in ASL parser on untyped
code.