Improvements to enable parsing and checking intermediate rewriting steps
Parser now has syntax for mutual recusion blocks

mutual {
  ... fundefs ...
}

which is used for parsing and pretty printing
DEF_internal_mutrec. It's stripped away by the initial_check, so the
typechecker never sees DEF_internal_mutrec. Maybe this could change,
as forcing mutual recursion to be explicit would probably be a good
thing.

Added record syntax to the new parser

New option -dmagic_hash is similar to GHC's -XMagicHash in that it
allows for identifiers to contain the special hash character, which is
used to introduce new autogenerated variables in a way that doesn't
clash with existing names.

Option -sil compiles sail down to the intermediate language defined in
sil.ott (not complete yet).