Un-tuple function arguments when pretty-printing to Lem
This makes Lem prefer plain "definition" when generating Isabelle output for
functions, which is processed by Isabelle much faster than "fun"/"function"
definitions.

This change is not complete yet, because the Sail library functions still need
to be changed to not use tupled arguments (possibly as part of a bigger
refactoring of the library).  For now, external functions are special-cased in
the pretty-printer and get tupled arguments.