Fix problem with let-bindings in pattern guards
Monomorphisation sometimes produces pattern guard with let-bindings, e.g.

| ... if (let regsize = size_itself(regsize) in eq(regsize, 32)) -> ...

Previously, the rewriting pass for let-bindings (and pattern guards) pulled
these out of the guard condition and into the same scope as the case
expression, which potentially clashed with let-bindings for the same variables
in that case expression.  The rewriter now leaves let-bindings in place within
pure if-conditions, solving this problem.