Rewrite topological sorting
Use Tarjan's algorithm for finding strongly connected components (and finding a
topological sorting of components at the same time), in order to properly take
into account mutually recursive functions.  The sorting is stable, i.e.,
definitions are only moved when necessary.  Exceptions to this are statements
that do not have any dependencies: default bitvector order declarations,
operator fixity declarations, and top-level comments.  These are moved to the
beginning (like with the previous sorting implementation).

Any dependency cycles that are found are additionally printed to the console in
dot-format, for easy visualisation with graphviz.