Try to remove early returns more aggressively
In particular, there is an ASL pattern with single-branch if-expressions
containing an early return (and an empty else-branch), e.g.

{
  ...
  if (error) then return(Fault) else ();
  ...
  return(Success);
}

The rewriting pass now tries to fold the rest of the block into the
else-branch, since it is unreachable after the then-branch, e.g.

{
  ...
  if (error) then return(Fault) else {
    ...
    return(Success);
  }
}

In combination with other rewriting rules, this allows the rewriting pass to
pull out the early returns if *all* branches end in a return statement.  If it
is possible to pull out the return all the way, i.e., so that the function body
is a single return statement, then the return can be removed. If that is not
possible, then the function body is left as it was originally.