Add simple conditional processing and file include
Can now use C-style include declarations to include files within other sail files. This is done in such a way that all the location information is preserved in error messages. As an example:

$include "aarch64/prelude.sail"
$define SYM
$ifndef SYM
$include <../util.sail>
$endif

would include the file aarch64/prelude.sail relative to the file where the include is contained. It then defines a symbol SYM and includes another file if it is not defined. The <../util.sail> include will be accessed relative to $SAIL_DIR/lib, so $SAIL_DIR/lib/../util.sail in this case.

This can be used with the standard C trick of

$ifndef ONCE
$define ONCE
val f : unit -> unit
$endif

so no matter how many sail files include the above file, the valspec for f will only appear once.

Currently we just have $include, $define, $ifdef and $ifndef (with $else and $endif). We're using $ rather than # because # is already used in internal identifiers, although this could be switched.