Add rreg effect to _reg_deref in fix_val_specs rewrite
The internal function _reg_deref is declared as pure, so that bitfield setters
can be implemented as read-modify-write, while only having a wreg effect.
However, for the Lem shallow embedding, the read step of those setters needs to
be embedded into the monad.  This could be special-cased in the Lem pretty
printer, but then the pretty printer would have to replicate some logic of the
letbind_effects rewriting step.  It seems simplest to add the effect annotation
early in the Lem rewriting pipeline, in the fix_val_specs step.  This means
that this rewriting step can only be used for other backends if these
additional effects are acceptable.