Comment out special casing of execute function in Lem pretty-printer
It assumes that execute is non-recursive, which is not the case for RISC-V with
compressed instructions.  Splitting execute into different auxiliary functions
for each clause is probably still useful, as Isabelle is likely to parse many
small functions faster than one big (potentially recursive) function, but this
splitting should be done in the rewriter instead of the pretty-printer, in
order to properly deal with recursion.