Add more annotations for loop bounds in Lem rewriting
Typechecking for-loops failed after the Lem rewriting passes in some cases: if
the lower bound for the loop may be greater than the upper bound, the loop
variable's type might be empty, and it cannot be initialised.  This patch adds
a guard "lower <= upper" around the loop body, and removes it again during
pretty-printing.