Main: exit if you hit IMPDEF behaviour
If you don't exit immediately, the test will probably just jump off into the
weeds and stay there until it times out.  So better to exit early.

But note that this is not all IMPDEF behaviour.
Most IMPDEF such as 'Are SHA3 crypto instructions available?' are already
being handled and either TRUE or FALSE is being returned.
So this is just for the stuff where we don't have a good answer at all.