RTS: Fail on AArch32 and ASIMD
This change causes execution of AArch32 code or ASIMD code
to fail with an easily diagnosed error message of the form

   UNIMPLEMENTED: xxx support

where xxx is either AArch32 or ASIMD.

Having a clear error message allows these to be detected by triaging
scripts.

To make the AArch32 detection part work, you also need to change
HaveAnyAArch32 to read like this instead of just returning false.

    function HaveAnyAArch32 () =
      return ((CFG_ID_AA64PFR0_EL1_EL0 == 0x2)
              || (CFG_ID_AA64PFR0_EL1_EL1 == 0x2)
              || (CFG_ID_AA64PFR0_EL1_EL2 == 0x2)
              || (CFG_ID_AA64PFR0_EL1_EL3 == 0x2))