An initial fix to riscv lr/sc, needs a review.
This uses a stronger model than the version in Sail-1 in order to
perform address alignment checks.  The reservation is kept on virtual
addresses, and maintained in the platform model, but now the lr/sc
definitions need calls to externs to update this state.  An
alternative was to reserve physical addresses, but that appeared to be
more complicated without a lot more changes.  Ideally, the model
should be parameterizable over both options.