Fix optimisation bug for certain if statements
When converting to A-normal form I just used the type of the then
branch of if statements to get the type of the whole if statement -
usually they'd be the same, but with flow typing one of the branches
can have a false constraint, which then allows the optimizer to fit
any integer into a 64-bit integer causing an overflow. The fix is to
correctly use the type the typechecker gives for the whole if
statement.

Also add decimal_string_of_bits to the C output.

Rename is_reftyp to is_ref_typ to be more consistent with other
is_X_typ functions in Ast_util.